# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exchange API",
        "description": "API documentation for ExchangeProject",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/expenses/export": {
            "get": {
                "summary": "Export expenses to Excel",
                "description": "Export all expenses to an Excel file.",
                "operationId": "1b3b26bd874fa2a93104a4e2b46a7972",
                "responses": {
                    "200": {
                        "description": "Excel file downloaded successfully."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/expenses": {
            "get": {
                "summary": "List all expenses for admin",
                "description": "Retrieve a paginated list of expenses for admin users.",
                "operationId": "36e62ecfbcfde37686ea63fe89d0c0c0",
                "responses": {
                    "200": {
                        "description": "Successful response."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/assistant/expenses": {
            "get": {
                "summary": "List all expenses for assistant",
                "description": "Retrieve a paginated list of expenses for assistant users.",
                "operationId": "4a2d3f8242495eb38aa91a8127b57659",
                "responses": {
                    "200": {
                        "description": "Successful response."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/expenses/{id}": {
            "delete": {
                "summary": "Delete an expense record",
                "description": "Delete a specific expense record by its ID.",
                "operationId": "b261a1443844b958952ae83b03fbbfa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted successfully."
                    },
                    "404": {
                        "description": "Expense not found."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/exchange/expenses": {
            "get": {
                "summary": "List all expenses for exchange users",
                "description": "Retrieve a paginated list of expenses for exchange users.",
                "operationId": "715f734f02663e0ab3214b87efb16533",
                "responses": {
                    "200": {
                        "description": "Successful response."
                    },
                    "401": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
...
    },
...
}